requirements.txt
# Testing framework
pytest>=7.0.0,<8.0.0

# HTTP client for API testing
requests>=2.28.0,<3.0.0

# Environment variable management
python-dotenv>=0.19.0,<2.0.0

# Date utilities for calendar calculations
python-dateutil>=2.8.0,<3.0.0

# HTTP retry mechanisms
urllib3>=1.26.0,<2.0.0